Task 0 convert to unsigned int

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -23,5 +23,5 @@ unsigned int binary_to_uint(const char *b)
 			return (0);
 		}
 	}
-	return (total);
+	return ((unsigned int) total);
 }
